+++ src/main/java/com/coredisc/domain/Comment.java
@@ -2,6 +2,7 @@ package com.coredisc.domain;
 
 import com.coredisc.domain.common.BaseEntity;
 import com.coredisc.domain.member.Member;
+import com.coredisc.domain.post.Post;
 import jakarta.persistence.*;
 import lombok.*;
