+++ src/main/java/com/coredisc/domain/PostAnswer.java
@@ -3,6 +3,7 @@ package com.coredisc.domain;
 
 import com.coredisc.domain.common.BaseEntity;
 import com.coredisc.domain.common.enums.AnswerType;
+import com.coredisc.domain.post.Post;
 import com.coredisc.domain.postAnswerImage.PostAnswerImage;
 import jakarta.persistence.*;
 import lombok.*;
