+++ src/main/java/com/coredisc/domain/member/Member.java
@@ -1,7 +1,6 @@
 package com.coredisc.domain.member;
 
 import com.coredisc.domain.Comment;
-import com.coredisc.domain.Post;
 import com.coredisc.domain.PostLike;
 import com.coredisc.domain.common.BaseEntity;
 import com.coredisc.domain.common.enums.OauthType;
@@ -9,6 +8,7 @@ import com.coredisc.domain.common.enums.Role;
 import com.coredisc.domain.follow.Follow;
 import com.coredisc.domain.mapping.MemberTerms;
 import com.coredisc.domain.monthlyReport.MonthlyReport;
+import com.coredisc.domain.post.Post;
 import com.coredisc.domain.profileImg.ProfileImg;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.persistence.*;
